+++ src/commands/MainCommand.ts
@@ -1,22 +1,23 @@
 import { Command, Option } from "clipanion"
 import GenerativeCommand from "../lib/GenerativeCommand"
+import ExplainCommand from "./ExplainCommand"
 
 export default class MainCommand extends GenerativeCommand {
   static usage = Command.Usage({
     description: "Generate content based on prompt",
   })
 
-  isShell = Option.Boolean("-s, --shell", { description: "Generate commands based on prompts" })
+  isExplain = Option.Boolean("-e, --explain", { description: ExplainCommand.usage.description })
   prompt = Option.Rest({ name: "prompt", required: 1 })
 
   async execute() {
     this.assertGeminiKey()
 
-    if (this.isShell) {
-      this.cli.run(["shell", ...this.prompt])
+    if (this.isExplain) {
+      this.cli.run(["explain", ...this.prompt])
       return
     }
 
-    this.cli.run(["get-assistance", ...this.prompt])
+    this.cli.run(["shell", ...this.prompt])
   }
 }