+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added custom instructions
+- Added explain command
+- Added keyboard shortcuts
+
+### Changed
+
+- Improved help
+
+### Fixed
+
+- Added support for pipe, redirects and backgrounding
+- Improved response generation accuracy
+
 ## [0.1.0] - 2024-04-03
 
 ### Added